+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next release
+
+### New features
+
+- Added `\Wizaplace\SDK\AuthLog\AuthLogService::get`
+- Added `\Wizaplace\SDK\AuthLog\AuthLogService::search`
+
 ## 1.48.0
 
 ### New features
@@ -35,7 +42,7 @@
 
 ## 1.44.13
 
- ### New features
+### New features
 
 - Added `\Wizaplace\SDK\Company\CompanyRegistration::getIban`
 - Added `\Wizaplace\SDK\Company\CompanyRegistration::setIban`
@@ -57,7 +64,7 @@
 ### Bugfixes
 
 - Fixed `\Wizaplace\SDK\Catalog\ProductOffer` with null `$data['divisions']`
-- Fixed `\Wizaplace\SDK\Catalog\ProductSummary::getOffers` return null 
+- Fixed `\Wizaplace\SDK\Catalog\ProductSummary::getOffers` return null
 
 ## 1.44.10